feat: create useAvatarGenerated
