+++ design-system/react/src/components/Avatar/AvatarGenerated.tsx
@@ -1,17 +1,10 @@
 import { cx } from "@fuel-ui/css";
-import { toSvg } from "jdenticon";
 
 import { createComponent } from "../../utils";
 import type { BoxProps } from "../Box";
 import { Box } from "../Box";
 
-import { getBackgroundColor } from "./utils";
-
-const SIZES = {
-  sm: 30,
-  md: 40,
-  lg: 50,
-};
+import { useAvatarGenerated } from "./hooks/useAvatarGenerated";
 
 type OmitProps = "children";
 export type AvatarGeneratedProps = BoxProps & {
@@ -26,13 +19,13 @@ export const AvatarGenerated = createComponent<
   unknown,
   OmitProps
 >(({ hash, size = "md", className, css, as, background, ...props }) => {
-  const totalSize = SIZES[size];
-  const backColor = getBackgroundColor(background, hash);
-  const svgString = toSvg(hash, totalSize, {
-    backColor,
-    padding: 0.15,
+  const { svgString, totalSize } = useAvatarGenerated({
+    background,
+    hash,
+    size,
   });
   const classes = cx(className, "fuel_avatar-generated");
+
   return (
     <Box
       {...props}
