+++ design-system/react/src/components/Avatar/hooks/useAvatarGenerated.tsx
@@ -1,0 +1,50 @@
+import { lightColors } from "@fuel-ui/css";
+import { toSvg } from "jdenticon";
+
+const SIZES = {
+  sm: 30,
+  md: 40,
+  lg: 50,
+};
+
+function getBackgroundColor(
+  backgroundColor?: string,
+  unique?: string
+): string | undefined {
+  if (backgroundColor === "fuel") return lightColors.brand;
+  if (backgroundColor !== "random") return backgroundColor;
+  if (!unique) return unique;
+
+  const hash = [...unique].reduce(
+    (prev, _cur, i) => unique.charCodeAt(i) + ((prev << 5) - prev),
+    0
+  );
+  return [0, 0, 0].reduce(
+    (prev, v, i) =>
+      `${prev}${`00${((hash >> (i * 8)) & 0xff).toString(16)}`.slice(-2)}`,
+    "#"
+  );
+}
+
+type UseAvatarGeneratedProps = {
+  background?: string;
+  hash?: string;
+  size: string;
+};
+
+export function useAvatarGenerated({
+  background,
+  hash,
+  size,
+}: UseAvatarGeneratedProps) {
+  const totalSize = SIZES[size];
+  const backColor = getBackgroundColor(background, hash);
+
+  return {
+    svgString: toSvg(hash, totalSize, {
+      backColor,
+      padding: 0.15,
+    }),
+    totalSize,
+  };
+}
